+++ context/index.ts
@@ -2,3 +2,4 @@ export * from "./request-context";
 export * from "./request-context.service";
 export * from "./request-context.interceptor";
 export * from "./global-response.interceptor";
+export * from "./context.module";